+++ src/lightning/pytorch/loops/prediction_loop.py
@@ -191,10 +191,7 @@ class _PredictionLoop(_Loop):
     def on_run_start(self) -> None:
         """Calls ``_on_predict_model_eval``, ``_on_predict_start`` and ``_on_predict_epoch_start`` hooks."""
         self._verify_dataloader_idx_requirement()
-
-        trainer = self.trainer
-        call._call_lightning_module_hook(trainer, "on_predict_model_eval")
-        trainer.lightning_module.zero_grad()
+        call._call_lightning_module_hook(self.trainer, "on_predict_model_eval")
         self._on_predict_start()
         self._on_predict_epoch_start()